ENH: Adding RCPS_CRC method in another file
