+++ mapie/multi_label_classification.py
@@ -20,6 +20,8 @@ from itertools import chain
 
 from .control_risk.ltt import find_lambda_control_star, ltt_procedure
 from .control_risk.risks import compute_risk_precision, compute_risk_recall
+from .control_risk.RCPS_CRC import _get_r_hat_plus
+from .control_risk.RCPS_CRC import _find_lambda_star
 
 
 class MapieMultiLabelClassifier(BaseEstimator, ClassifierMixin):
@@ -102,9 +104,9 @@ class MapieMultiLabelClassifier(BaseEstimator, ClassifierMixin):
 
     valid_index: List[List[Any]]
         List of list of all index that satisfy fwer controlling
-        for learn then test procedure. This attribute is compute
-        when the user wants to control precision score. Contain
-        a number of n_alpha lists (see predict).
+        for learn then test procedure. This attribute is computed
+        when the user wants to control precision score. Contains
+        n_alpha lists (see predict).
 
     References
     ----------
@@ -146,7 +148,6 @@ class MapieMultiLabelClassifier(BaseEstimator, ClassifierMixin):
         "single_estimator_",
         "risks"
     ]
-    sigma_init = .25  # Value given in the paper.
     cal_size = .3
 
     def __init__(
@@ -255,8 +256,8 @@ class MapieMultiLabelClassifier(BaseEstimator, ClassifierMixin):
             if delta is None:
                 raise ValueError(
                     "Invalid delta. "
-                    "delta cannot be ``None`` when using "
-                    "RCPS or LTT method."
+                    "delta cannot be ``None`` when controlling "
+                    "Recall with RCPS or Precision with LTT"
                 )
             elif ((delta <= 0) or (delta >= 1)):
                 raise ValueError(
@@ -469,170 +470,6 @@ class MapieMultiLabelClassifier(BaseEstimator, ClassifierMixin):
 
         return np.expand_dims(y_pred_proba_array, axis=2)
 
-    def _get_r_hat_plus(
-        self,
-        bound: Optional[str],
-        delta: Optional[float],
-    ) -> Tuple[NDArray, NDArray]:
-        """Compute the upper bound of the loss for each lambda.
-
-        Parameters
-        ----------
-        bound : str
-            Bounds to compute. Either hoeffding, bernstein or wsr.
-        delta : float
-            Level of confidence.
-        sigma_init : float, optional
-            First variance in the sigma_hat array. The default
-            value is the same as in the paper implementation.
-
-            By default .25
-
-        Returns
-        -------
-        Tuple[NDArray, NDArray] of shape (n_lambdas, ) and (n_lambdas)
-            Average risk over all the obervations and upper bound of the risk.
-        """
-        r_hat = self.risks.mean(axis=0)
-        n_obs = len(self.risks)
-
-        if (self.method == "rcps") and (delta is not None):
-            if bound == "hoeffding":
-                r_hat_plus = (
-                    r_hat +
-                    np.sqrt((1 / (2 * n_obs)) * np.log(1 / delta))
-                )
-
-            elif bound == "bernstein":
-                sigma_hat_bern = np.var(r_hat, axis=0, ddof=1)
-                r_hat_plus = (
-                    r_hat +
-                    np.sqrt((sigma_hat_bern * 2 * np.log(2 / delta)) / n_obs) +
-                    (7 * np.log(2 / delta)) / (3 * (n_obs - 1))
-                )
-
-            else:
-                mu_hat = (
-                    (.5 + np.cumsum(self.risks, axis=0)) /
-                    (np.repeat(
-                        [range(1, n_obs + 1)],
-                        self.n_lambdas,
-                        axis=0
-                    ).T + 1)
-                )
-                sigma_hat = (
-                    (.25 + np.cumsum((self.risks - mu_hat)**2, axis=0)) /
-                    (np.repeat(
-                        [range(1, n_obs + 1)],
-                        self.n_lambdas,
-                        axis=0
-                    ).T + 1)
-                )
-                sigma_hat = np.concatenate(
-                    [
-                        np.full(
-                            (1, self.n_lambdas), fill_value=self.sigma_init
-                        ), sigma_hat[:-1]
-                    ]
-                )
-                nu = np.minimum(
-                    1,
-                    np.sqrt((2 * np.log(1 / delta)) / (n_obs * sigma_hat))
-                )
-
-                # Split the calculation in two to prevent memory issues
-                batches = [
-                    range(int(n_obs / 2)),
-                    range(n_obs - int(n_obs / 2), n_obs)
-                ]
-                K_R_max = np.zeros((self.n_lambdas, self.n_lambdas))
-                for batch in batches:
-                    nu_batch = nu[batch]
-                    losses_batch = self.risks[batch]
-
-                    nu_batch = np.repeat(
-                        np.expand_dims(nu_batch, axis=2),
-                        self.n_lambdas,
-                        axis=2
-                    )
-                    losses_batch = np.repeat(
-                        np.expand_dims(losses_batch, axis=2),
-                        self.n_lambdas,
-                        axis=2
-                    )
-
-                    R = self.lambdas
-                    K_R = np.cumsum(
-                        np.log(
-                            (
-                                1 -
-                                nu_batch *
-                                (losses_batch - R)
-                            ) +
-                            np.finfo(np.float64).eps
-                        ),
-                        axis=0
-                    )
-                    K_R = np.max(K_R, axis=0)
-                    K_R_max += K_R
-
-                r_hat_plus_tronc = self.lambdas[np.argwhere(
-                    np.cumsum(K_R_max > -np.log(delta), axis=1) == 1
-                )[:, 1]]
-                r_hat_plus = np.ones(self.n_lambdas)
-                r_hat_plus[:len(r_hat_plus_tronc)] = r_hat_plus_tronc
-
-        else:
-            r_hat_plus = (n_obs / (n_obs + 1)) * r_hat + (1 / (n_obs + 1))
-
-        return r_hat, r_hat_plus
-
-    def _find_lambda_star(
-        self,
-        r_hat_plus: NDArray,
-        alpha_np: NDArray
-    ) -> NDArray:
-        """Find the higher value of lambda such that for
-        all smaller lambda, the risk is smaller, for each value
-        of alpha.
-
-        Parameters
-        ----------
-        r_hat_plus : NDArray of shape (n_lambdas, )
-            Upper bounds computed in the `get_r_hat_plus` method.
-        alphas : NDArray of shape (n_alphas, )
-            Risk levels.
-
-        Returns
-        -------
-        NDArray of shape (n_alphas, )
-            Optimal lambdas which control the risks for each value
-            of alpha.
-        """
-
-        if len(alpha_np) > 1:
-            alphas_np = alpha_np[:, np.newaxis]
-        else:
-            alphas_np = alpha_np
-
-        bound_rep = np.repeat(
-            np.expand_dims(r_hat_plus, axis=0),
-            len(alphas_np),
-            axis=0
-        )
-        bound_rep[:, np.argmax(bound_rep, axis=1)] = np.maximum(
-            alphas_np,
-            bound_rep[:, np.argmax(bound_rep, axis=1)]
-        )  # to avoid an error if the risk is always higher than alpha
-        lambdas_star = self.lambdas[np.argmin(
-                - np.greater_equal(
-                    bound_rep,
-                    alphas_np
-                ).astype(int),
-                axis=1
-            )]
-
-        return lambdas_star
 
     def partial_fit(
         self,
@@ -839,11 +676,12 @@ class MapieMultiLabelClassifier(BaseEstimator, ClassifierMixin):
             )
 
         else:
-            self.r_hat, self.r_hat_plus = self._get_r_hat_plus(
-                bound, delta,
+            self.r_hat, self.r_hat_plus = _get_r_hat_plus(
+                self.risks, self.lambdas, self.method,
+                bound, delta
             )
-            self.lambdas_star = self._find_lambda_star(
-                self.r_hat_plus, alpha_np
+            self.lambdas_star = _find_lambda_star(
+                self.lambdas, self.r_hat_plus, alpha_np
             )
             y_pred_proba_array = (
                 y_pred_proba_array >
